consolidate and internalize utilities
